Remove obsolete Router
