+++ lib/openapi_first.rb
@@ -6,7 +6,6 @@ require_relative 'openapi_first/configuration'
 require_relative 'openapi_first/plugins'
 require_relative 'openapi_first/definition'
 require_relative 'openapi_first/version'
-require_relative 'openapi_first/router'
 require_relative 'openapi_first/error_response'
 require_relative 'openapi_first/request_validation'
 require_relative 'openapi_first/response_validation'
