+++ lib/openapi_first/definition/path_parameters.rb
@@ -2,7 +2,6 @@
 
 require 'openapi_parameters'
 require_relative 'parameters'
-require_relative '../router'
 
 module OpenapiFirst
   class PathParameters < Parameters
